Move mode to typed_terms
